requirements.txt
# --- Core / ML ---
torch
torchvision
numpy
torchmetrics
scipy

# --- Image & Data Processing ---
opencv-python
Pillow
albumentations
scikit-image

# --- Metrics ---
piq

# --- GPU / Acceleration ---
onnx
onnxruntime-gpu
cupy-cuda12x

# --- C++ / Performance ---
pybind11
cmake

# --- Experiment Tracking ---
mlflow
wandb

# --- Data Versioning ---
dvc

# --- Config / Utils ---
pydantic
pydantic-settings
pyyaml
python-dotenv
hydra-core

# --- API / Backend ---
fastapi
uvicorn
python-multipart

# --- Monitoring / Profiling ---
prometheus-client
psutil
pynvml
line_profiler
memory_profiler

# --- Testing ---
pytest
pytest-benchmark
pytest-cov

# --- Dev / Code Quality ---
black
isort
flake8
mypy

# --- General Utilities ---
tqdm
loguru
rich
huggingface-hub

# --- GCP Storage ---
google-cloud-storage

# --- Database ---
sqlalchemy
alembic
psycopg2-binary

# --- Kafka ---
confluent-kafka

# --- IDs / Utils ---
python-magic